reduce, h1
WAY too much! but that's okay! now we know that margin is all the way around the outside, as if the whole page is a box, and padding is around each tag (?): the effect here was space between all the lines, but they are all different heading types, so we'll see